+++ src/angular/planit/src/app/map/map-routing.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { ExpirationGuard } from '../core/services/expiration-guard.service';
+import { PlanAuthGuard } from '../core/services/plan-auth-guard.service';
+import { MapComponent } from './map.component';
+
+const routes: Routes = [
+  {
+    path: 'map',
+    component: MapComponent,
+    canActivate: [ExpirationGuard, PlanAuthGuard]
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class MapRoutingModule { }