+++ src/angular/planit/src/app/dashboard/city-minimap/city-minimap.component.ts
@@ -1,0 +1,17 @@
+import { Component, Input } from '@angular/core';
+
+import { Observable } from 'rxjs';
+
+import { Organization } from '../../shared/';
+
+@Component({
+  selector: 'app-city-minimap',
+  templateUrl: 'city-minimap.component.html'
+})
+
+export class CityMinimapComponent {
+
+  @Input() public organization: Organization;
+
+  constructor() { }
+}