+++ deployment/terraform/variables.tf
@@ -64,6 +64,10 @@ variable "ccapi_host" {}
 
 variable "planit_app_home" {}
 
+variable "esri_client_id" {}
+
+variable "esri_client_secret" {}
+
 variable "ecs_cloudwatch_events_role" {
   default = "arn:aws:iam::aws:policy/service-role/AmazonEC2ContainerServiceEventsRole"
 }