+++ src/angular/planit/src/app/core/interceptors/AuthInterceptor.ts
@@ -5,6 +5,7 @@ import { Injectable } from '@angular/core';
 
 import { Observable } from 'rxjs';
 import { AuthService } from '../services/auth.service';
+import { environment } from '../../../environments/environment';
 
 /** Pass untouched request through to the next request handler. */
 @Injectable()
@@ -13,7 +14,7 @@ export class AuthInterceptor implements HttpInterceptor {
   constructor(private authService: AuthService) {}
 
   intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-    if (!this.authService.isAuthenticated()) {
+    if (!this.authService.isAuthenticated() || !req.url.startsWith(environment.apiUrl)) {
       return next.handle(req);
     }