+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Display related impacts in Risk wizard & vulnerability assessment overview
  - Switched from Google Places to ESRI geocoder
  - Added 'geocode_organizations' management command
- - Added tribal areas layer to the organization area map
+ - Added tribal areas and county layers to the organization area map
 ### Fixed
  - Fixed city profile missing from dashboard