+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!--When making a new release, remember to update the magic links at the bottom.-->
-## [Unreleased]
+## [1.11.0] - 2020-01-09
 ### Added
  - Updated Plan Wizard to allow selecting Organization boundaries
  - Added map page
@@ -168,7 +168,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/azavea/temperate/compare/1.9.0...HEAD
+[Unreleased]: https://github.com/azavea/temperate/compare/1.11.0...HEAD
+[1.11.0]: https://github.com/azavea/temperate/compare/1.10.0...1.11.0
 [1.10.0]: https://github.com/azavea/temperate/compare/1.9.0...1.10.0
 [1.9.0]: https://github.com/azavea/temperate/compare/1.8.0...1.9.0
 [1.8.0]: https://github.com/azavea/temperate/compare/1.7.0...1.8.0